Renaming singing decorator (#539)
